test(mitm-addon): split registry test suite
